+++ src/java/servlets/serv_DeleteStudent.java
@@ -16,17 +16,16 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
-
 import network.Login;
 
 /**
  *
  * @author Tobias
  */
-@WebServlet(name = "deleteModule", urlPatterns = {"/deleteModule"})
-public class deleteModule extends HttpServlet {
+@WebServlet(name = "deleteStudent", urlPatterns = {"/deleteStudent"})
+public class serv_DeleteStudent extends HttpServlet {
     Login login = new Login();
-
+    
     /**
      * Processes requests for both HTTP <code>GET</code> and <code>POST</code>
      * methods.
@@ -42,25 +41,26 @@ public class deleteModule extends HttpServlet {
         try (PrintWriter out = response.getWriter()) {
             
             HtmlHelper site = new HtmlHelper(out);
-            site.printHead("Delete module", "delete-module");
+            site.printHead("Delete student", "delete-student");
             
             out.println("<h1>Servlet deleteStudent at " + request.getContextPath() + "</h1>");
             
             Connection conn = login.loginToDB(out);
             
-            String module_id = request.getParameter("module_id");
+            String student_id = request.getParameter("student_id");
             
-            PreparedStatement deleteModule;
+            PreparedStatement deleteStudent;
             try {
-                deleteModule = conn.prepareStatement("DELETE FROM module WHERE mod_id = ?;");
-                deleteModule.setString(1, module_id);
+                deleteStudent = conn.prepareStatement("DELETE FROM student WHERE student_id = ?;");
+                deleteStudent.setString(1, student_id);
                 
-                int amountDeleted = deleteModule.executeUpdate();
-                out.println("<div>" + amountDeleted + " modules deleted.</div>");
-                out.println("<a href=\"getModule\">Back to module list</a>");
+                int amountDeleted = deleteStudent.executeUpdate();
+                out.println("<div>" + amountDeleted + " students deleted.</div>");
+                out.println("<a href=\"getStudent\">Back to student list</a>");
             } catch (SQLException ex) {
                 out.println("SQL error: " + ex);
             }
+            
             
             site.printEnd();
         }
