+++ src/java/servlets/serv_OneStudent.java
@@ -5,32 +5,29 @@
  */
 package servlets;
 
-import helpers.HtmlHelper;
-import helpers.StudentHelper;
 import java.io.IOException;
 import java.io.PrintWriter;
-import java.sql.Connection;
-import java.sql.PreparedStatement;
-import java.sql.ResultSet;
-import java.sql.SQLException;
-import java.sql.Statement;
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+
+import helpers.*;
+import java.sql.Connection;
+import java.sql.Statement;
 import network.Login;
 
 /**
  *
- * @author Staven
+ * @author Tobias
  */
-@WebServlet(name = "getStudent", urlPatterns = {"/getStudent"})
-public class getStudent extends HttpServlet {
-
+@WebServlet(name = "oneStudent", urlPatterns = {"/oneStudent"})
+public class serv_OneStudent extends HttpServlet {
+    
     Statement stmt;
     Login login = new Login();
-    
+
     /**
      * Processes requests for both HTTP <code>GET</code> and <code>POST</code>
      * methods.
@@ -44,20 +41,23 @@ public class getStudent extends HttpServlet {
             throws ServletException, IOException {
         response.setContentType("text/html;charset=UTF-8");
         try (PrintWriter out = response.getWriter()) {
-            
+            /* TODO output your page here. You may use following sample code. */
             HtmlHelper site = new HtmlHelper(out);
-            site.printHead("Students", "bodyy");
+            site.printHead("Single student", "one-student-container");
             
-            out.println("<h1>Servlet getStudent at " + request.getContextPath() + "</h1>");
+            String stid = request.getParameter("stid");
+            
+            Connection conn;
+            conn = login.loginToDB(out);
 
-                Connection conn;
-                conn = login.loginToDB(out);
-                
-                StudentHelper.printStudents(out, conn);
-                
-                
-                login.close();
-                
+            
+            out.println("<h2>Viewing a single student</h2>");
+            StudentHelper.printOneStudent(out, conn, stid);
+
+            
+            login.close();
+            
+            
             site.printEnd();
         }
     }
@@ -74,9 +74,6 @@ public class getStudent extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
-        response.setContentType("text/html;charset=UTF-8");
-        response.setCharacterEncoding("UTF-8");
-        request.setCharacterEncoding("UTF-8");  
         processRequest(request, response);
     }
 
@@ -91,9 +88,6 @@ public class getStudent extends HttpServlet {
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
-        response.setContentType("text/html;charset=UTF-8");
-        response.setCharacterEncoding("UTF-8");
-        request.setCharacterEncoding("UTF-8"); 
         processRequest(request, response);
     }
