+++ src/java/servlets/serv_UpdateModule.java
@@ -6,7 +6,7 @@
 package servlets;
 
 import helpers.HtmlHelper;
-import helpers.StudentHelper;
+import helpers.ModuleHelper;
 import java.io.IOException;
 import java.io.PrintWriter;
 import java.sql.Connection;
@@ -19,13 +19,12 @@ import network.Login;
 
 /**
  *
- * @author Staven
+ * @author tobia
  */
-@WebServlet(name = "createStudent", urlPatterns = {"/createStudent"})
-public class createStudent extends HttpServlet {
-
+@WebServlet(name = "updateModule", urlPatterns = {"/updateModule"})
+public class serv_UpdateModule extends HttpServlet {
     Login login = new Login();
-    
+
     /**
      * Processes requests for both HTTP <code>GET</code> and <code>POST</code>
      * methods.
@@ -39,22 +38,30 @@ public class createStudent extends HttpServlet {
             throws ServletException, IOException {
         response.setContentType("text/html;charset=UTF-8");
         try (PrintWriter out = response.getWriter()) {
-            
             HtmlHelper site = new HtmlHelper(out);
-            site.printHead("New student", "create-student");
             
-                Connection conn;
-                conn = login.loginToDB(out);
-                
-                StudentHelper.insertStudent(
-                        request.getParameter("student_name"),
-                        request.getParameter("student_edu"),
-                        conn, 
-                        out
-                );
-                
-                login.close();
-                
+            //body class 'invisible' makes no content on the page visible, and it should auto load
+            //due to javascript 
+            site.printHead("Updating module...", "invisible");
+            
+            out.println("<h1>Servlet updateModule at " + request.getContextPath() + "</h1>");
+            
+            String id = request.getParameter("singleMod_id");
+            String name = request.getParameter("mod_name");
+            String desc = request.getParameter("mod_desc");
+            
+            Connection conn = login.loginToDB(out);
+            ModuleHelper.updateModule(id, name, desc, conn, out);
+            
+            //form that takes you back to the module you just edited
+            out.println("<form name=\"auto\" action=\"oneModule\">");
+            out.println("<input name=\"singleMod_id\" type=\"text\" value=\"" + id + "\">");
+            out.println("<input type=\"submit\">");
+            out.println("</form>");
+            
+            //auto submits the form so the page auto loads
+            out.println("<script>window.onload=document.forms[\'auto\'].submit();</script>");
+            
             site.printEnd();
         }
     }
@@ -71,9 +78,6 @@ public class createStudent extends HttpServlet {
     @Override
     protected void doGet(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
-        response.setContentType("text/html;charset=UTF-8");
-        response.setCharacterEncoding("UTF-8");
-        request.setCharacterEncoding("UTF-8"); 
         processRequest(request, response);
     }
 
@@ -88,9 +92,6 @@ public class createStudent extends HttpServlet {
     @Override
     protected void doPost(HttpServletRequest request, HttpServletResponse response)
             throws ServletException, IOException {
-        response.setContentType("text/html;charset=UTF-8");
-        response.setCharacterEncoding("UTF-8");
-        request.setCharacterEncoding("UTF-8"); 
         processRequest(request, response);
     }
